# api/openapi.yaml
openapi: 3.0.3
info:
  title: Dungeon Store API
  description: A basic API to buy minion and store them in dungeons.
  version: 1.0.0
servers:
- url: /
paths:
...
  /cart/{cartId}:
    summary: Specific cart using cartId
    description: This path is used for interacting with a specific cart using the
      cartId
    get:
      tags:
      - Cart
      summary: Get cart
      description: This path is used to get cart and it's minions
      operationId: get_cart_cart_id
      parameters:
      - name: cartId
        in: path
        description: The id of the dungeon.
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        default:
          description: Default error sample response
      x-swagger-router-controller: Cart
    post:
      tags:
      - Cart
      summary: Add minion to cart
      description: This path is used to add minions to the cart
      operationId: post_cart_cart_id
      parameters:
      - name: cartId
        in: path
        description: The id of the dungeon.
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        default:
          description: Default error sample response
      x-swagger-router-controller: Cart
...